NaverAPI 변경에 따라 코드 수정 및 개선

+++ Relay/src/relay/controller/wishlist/WishlistController.java
@@ -50,8 +50,8 @@ public class WishlistController {
 	}
 	
 	@RequestMapping("/wishlist/cnt.do")
-	public int getTotal(@RequestParam("id") String memberId) {
-		int cnt = wBiz.getTotal(memberId);
+	public long getTotal(@RequestParam("id") String memberId) {
+		long cnt = wBiz.getTotal(memberId);
 		return cnt;
 	}
